requirements.txt
requests==2.20
SQLAlchemy==1.2.12
mysqlclient==1.3.13
Flask==1.0.0
python-dotenv==0.10.0
timezonefinder==3.4.2
pytz ==2018.9
python-dotenv==0.10.0